POP 0.3.1.2
Breaking change:
POP now doesn't send a response if the command is a comment

Discussion:
The last commit introduced a bug where POP would send a response even if the command is a comment. This behavior would unnecessarily flood the pipe and QOQ was already configured to ignore a response for a comment.

POP:
- Changed PopServer to don't send a response if the command is a comment